backend/requirements.txt
# Django Core
Django==5.2.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
django-environ==0.11.2

# Security & Authentication
django-oauth-toolkit==2.3.0
django-ratelimit==4.1.0

# Image Processing & Biometrics
# Keep install light for defense/demo. Heavy libs (dlib/face-recognition) are optional.
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
# Optional (install if you plan to implement native face recognition)
# face-recognition==1.3.0
# dlib==19.24.2

# API & Serialization
drf-yasg==1.21.7
django-filter==24.1

# Background Tasks
celery==5.3.4
redis==5.0.1

# Monitoring & Logging
django-debug-toolbar==4.3.0
sentry-sdk==1.40.4

# Development & Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.4.1

# Production
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
